auth.gethomepage

+++ src/genres/genres.controller.ts
@@ -1,10 +1,12 @@
-import { Controller, Get, Post, Patch, Delete, Body, Param, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Patch, Delete, Body, Param, UseGuards, UnauthorizedException } from '@nestjs/common';
 import { GenresService } from './genres.service';
 import { CreateGenresDto } from './dto/CreateGenres.dto';
 import { UpdateGenresDto } from './dto/UpdateGenres.dto';
 import { ApiTags, ApiOperation, ApiBearerAuth } from '@nestjs/swagger';
-import { Genres } from '@prisma/client';
 import { AuthGuard } from '@nestjs/passport';
+import { LoggedUser } from '../auth/logged-user.decorator';
+import { Users } from 'src/users/entities/user.entity';
+import { Genres } from './entities/genres.entity';
 
 @ApiTags('genres')
 @UseGuards(AuthGuard())
@@ -14,36 +16,51 @@ export class GenresController {
     constructor(private readonly genresService: GenresService) {}
 
     @Get()
-    findAll() {
-        return this.genresService.findAll();
+    findAll(@LoggedUser() user: Users) {
+      if( !user.isAdmin ) {
+        throw new UnauthorizedException(`Usuário ${user.name} não esta cadastrado como administrador`)
+      }
+      return this.genresService.findAll();
     }
 
     @Get(':id')
     @ApiOperation({
         summary: 'Localizar um gênero pelo ID',
     })    
-    findOne(@Param('id') id: string) {
+    findOne(@Param('id') id: string, @LoggedUser() user: Users) {
+      if( !user.isAdmin ) {
+        throw new UnauthorizedException(`Usuário ${user.name} não esta cadastrado como administrador`)
+      }
       return this.genresService.findById(+id);
     }
 
     @Post()
-    create(@Body() dto: CreateGenresDto) {
-        return this.genresService.create(dto);
+    create(@Body() dto: CreateGenresDto, @LoggedUser() user: Users) {
+      if( !user.isAdmin ) {
+        throw new UnauthorizedException(`Usuário ${user.name} não esta cadastrado como administrador`)
+      }
+      return this.genresService.create(dto);
     }    
 
     @Patch(':id')
     @ApiOperation({
       summary: 'Editar um gênero pelo ID',
     })    
-    update(@Param('id') id: string, @Body() dto: UpdateGenresDto): Promise<Genres> {
-        return this.genresService.update( +id, dto);
+    update(@Param('id') id: string, @LoggedUser() user: Users ,@Body() dto: UpdateGenresDto): Promise<Genres> {
+      if( !user.isAdmin ) {
+        throw new UnauthorizedException(`Usuário ${user.name} não esta cadastrado como administrador`)
+      }
+      return this.genresService.update( +id, dto);
     }
 
     @Delete(':id')
     @ApiOperation({
       summary: 'Deletar gênero pelo ID',
     })    
-    delete(@Param('id') id: string,): Promise<Genres> {
-        return this.genresService.delete( +id );
+    delete(@Param('id') id: string, @LoggedUser() user: Users): Promise<Genres> {
+      if( !user.isAdmin ) {
+        throw new UnauthorizedException(`Usuário ${user.name} não esta cadastrado como administrador`)
+      }
+      return this.genresService.delete( +id );
     }
 }
